+++ ImageFeed/Models/ProfileImage.swift
@@ -1,0 +1,14 @@
+//
+//  ProfileImage.swift
+//  ImageFeed
+//
+//  Created by Owi Lover on 10/3/24.
+//
+
+import Foundation
+
+struct ProfileImage: Codable {
+    let small: String?
+    let medium: String?
+    let large: String?
+}
