+++ ImageFeed/Models/UserResult.swift
@@ -1,0 +1,12 @@
+//
+//  UserResult.swift
+//  ImageFeed
+//
+//  Created by Owi Lover on 10/3/24.
+//
+
+import Foundation
+
+struct UserResult: Codable {
+    let profileImage: ProfileImage
+}
